+++ quartz/components/Footer.tsx
@@ -13,10 +13,7 @@ export default ((opts?: Options) => {
     const links = opts?.links ?? []
     return (
       <footer class={`${displayClass ?? ""}`}>
-        <p>
-          {i18n(cfg.locale).components.footer.createdWith}{" "}
-          <a href="https://quartz.jzhao.xyz/">Quartz v{version}</a> © {year}
-        </p>
+        <p>路虽远，行则将至</p>
         <ul>
           {Object.entries(links).map(([text, link]) => (
             <li>